Organize romanization tables
